requirements.txt
fastapi==0.115.12
uvicorn==0.29.0
pydantic==2.11.3
pydantic_core==2.33.1
sqlalchemy==2.0.28
firebase-admin==6.6.0
alembic==1.13.1
python-dotenv==0.19.0
email_validator==2.2.0
requests==2.32.3
python-multipart==0.0.5
cryptography==44.0.0
pytz==2024.2
psycopg2-binary==2.9.9  # For PostgreSQL support
PyJWT==2.10.1
python-jose==3.3.0
passlib==1.7.4
loguru==0.7.3
httpx==0.27.2
pytest==8.3.5
reportlab==4.4.0
selenium==4.32.0
webdriver-manager==4.0.1

# Production dependencies
gunicorn==23.0.0
# uvloop==0.19.0 # Removed because it doesn't support Windows
httptools==0.6.4
prometheus-client==0.20.0
sentry-sdk==2.20.0
pg8000==1.29.8  # Pure-Python PostgreSQL driver (alternative to psycopg2)